null safety (not checked 100%)

+++ lib/utils/number.dart
@@ -1,10 +1,8 @@
-
 import 'dart:math';
 
 class NumberUtils {
-
   static double toPrecision(double number, int fractionDigits) {
-    double mod = pow(10, fractionDigits.toDouble());
+    double mod = pow(10, fractionDigits.toDouble()).toDouble();
     return ((number * mod).round().toDouble() / mod);
   }
 }
